feat: searchApi 생성 및 타입 파일 이동
